[FTR][SLOT-158]: Add search for students in a specific slot. Add invalidate tags for userCalendar when we modify a general slot
